minor fixes in type cast, deleted libft also, just becouse

+++ ft_print_hex.c
@@ -6,7 +6,7 @@
 /*   By: jomiguel < jomiguel@student.42sp.org.br    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/09 15:43:24 by jomiguel          #+#    #+#             */
-/*   Updated: 2022/05/09 15:45:58 by jomiguel         ###   ########.fr       */
+/*   Updated: 2022/05/09 18:39:23 by jomiguel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int	ft_print_hex(unsigned int ui)
 	char	*s;
 	int		n;
 
-	s = ft_uitoa_base(ui, HEXADECIMAL_BASE);
+	s = ft_uitoa_base((unsigned long)ui, HEXADECIMAL_BASE);
 	n = ft_print_string(s);
 	free(s);
 	return (n);
